Implemented BoolQuery, FilterQuery. Deprecated Bool
