+++ DSL/Bool/Bool.php
@@ -13,33 +13,15 @@ namespace ONGR\ElasticsearchBundle\DSL\Bool;
 
 use ONGR\ElasticsearchBundle\DSL\BuilderInterface;
 use ONGR\ElasticsearchBundle\DSL\ParametersTrait;
+use ONGR\ElasticsearchBundle\DSL\Query\BoolQuery;
 
 /**
  * Bool operator. Can be used for filters and queries.
+ *
+ * @deprecated Will be removed in 1.0. Use ONGR\ElasticsearchBundle\DSL\Query\BoolQuery.
  */
-class Bool implements BuilderInterface
+class Bool extends BoolQuery
 {
-    use ParametersTrait;
-
-    const MUST = 'must';
-    const MUST_NOT = 'must_not';
-    const SHOULD = 'should';
-
-    /**
-     * @var array
-     */
-    private $container = [];
-
-    /**
-     * Checks if bool filter is relevant.
-     *
-     * @return bool
-     */
-    public function isRelevant()
-    {
-        return (bool)count($this->container);
-    }
-
     /**
      * Add BuilderInterface object to bool operator.
      *
@@ -47,40 +29,11 @@ class Bool implements BuilderInterface
      * @param string           $type
      *
      * @throws \UnexpectedValueException
+     *
+     * @deprecated Will be removed in 1.0. Use ONGR\ElasticsearchBundle\DSL\Query\BoolQuery::add().
      */
-    public function addToBool(BuilderInterface $bool, $type = self::MUST)
+    public function addToBool(BuilderInterface $bool, $type = BoolQuery::MUST)
     {
-        $constants = (new \ReflectionObject($this))->getConstants();
-
-        if (!in_array($type, $constants)) {
-            throw new \UnexpectedValueException(sprintf('The bool operator %s is not supported', $type));
-        }
-
-        $this->container[$type][] = $bool;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getType()
-    {
-        return 'bool';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function toArray()
-    {
-        $output = $this->processArray();
-
-        foreach ($this->container as $type => $filters) {
-            /** @var BuilderInterface $bool */
-            foreach ($filters as $bool) {
-                $output[$type][] = [$bool->getType() => $bool->toArray()];
-            }
-        }
-
-        return $output;
+        $this->add($bool, $type);
     }
 }
