+++ DSL/Filter/BoolFilter.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the ONGR package.
+ *
+ * (c) NFQ Technologies UAB <info@nfq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace ONGR\ElasticsearchBundle\DSL\Filter;
+
+use ONGR\ElasticsearchBundle\DSL\Query\BoolQuery;
+
+/**
+ * Represents Elasticsearch "bool" filter.
+ *
+ * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-bool-filter.html
+ */
+class BoolFilter extends BoolQuery
+{
+}
